requirements.txt
drf-yasg>=1.21.4
django
djangorestframework